requirements.txt
psutil>=5.9.0
flask>=2.0.0
flask-cors>=3.0.0
requests>=2.26.0
pyobjc-framework-Cocoa>=9.0.1
pyobjc-framework-Quartz>=9.0.1
python-dotenv>=0.19.0
cryptography>=41.0.0
openai>=1.0.0
selenium>=4.0.0
webdriver_manager>=3.8.0 